+++ 04-helper/src/main/scala/orchescala/helper/dev/update/GenericFileGenerator.scala
@@ -1,5 +1,7 @@
 package orchescala.helper.dev.update
 
+import orchescala.helper.util.PipelineConfig
+
 case class GenericFileGenerator()(using config: DevConfig):
 
   lazy val generate: Unit =
@@ -12,6 +14,8 @@ case class GenericFileGenerator()(using config: DevConfig):
     os.makeDir.all(config.projectDir / ".vscode")
     createOrUpdate(config.projectDir / ".run" / "WorkerTestApp.run.xml", workerTestAppIntellij)
     createOrUpdate(config.projectDir / ".vscode" / "launch.json", workerTestAppVsCode)
+    config.pipelineConfig.foreach: pConfig => // only if configured
+      createOrUpdate(config.projectDir / ".gitlab-ci.yml", gitLabPipeline(pConfig))
   end generate
 
   lazy val generateForGateway: Unit =
@@ -141,6 +145,43 @@ case class GenericFileGenerator()(using config: DevConfig):
         |  </configuration>
         |</component>
         |""".stripMargin
+
+  private def gitLabPipeline(pipelineConfig: PipelineConfig) =
+    s"""
+       |# $helperDoNotAdjustText
+       |stages:
+       |  - test
+       |
+       |include:
+       |  - template: Jobs/SAST.gitlab-ci.yml
+       |  - template: Jobs/SAST-IaC.gitlab-ci.yml
+       |  - template: Jobs/Dependency-Scanning.gitlab-ci.yml
+       |  - template: Jobs/Secret-Detection.gitlab-ci.yml
+       |
+       |variables:
+       |  ALL_PROXY: ${pipelineConfig.baseProxy}
+       |  TP_PROXY: $$ALL_PROXY
+       |  HTTP_PROXY: $$ALL_PROXY
+       |  HTTPS_PROXY: $$ALL_PROXY
+       |  SCALA_IMAGE: ${pipelineConfig.baseImage}
+       |  ${pipelineConfig.companyMVNUser}: $$CI_REGISTRY_USER
+       |  ${pipelineConfig.companyMVNPassword}: $$CI_REGISTRY_PASSWORD
+       |
+       |worker-test:
+       |  stage: test
+       |  image:
+       |    name: $$SCALA_IMAGE
+       |  variables:
+       |    CI_DEBUG_SERVICES: false
+       |  script:
+       |    - echo $$CI_REGISTRY_USER
+       |    - curl -fL "https://github.com/coursier/launchers/raw/master/cs-x86_64-pc-linux.gz" | gzip -d > cs
+       |    - chmod +x ./cs
+       |    - eval "$$(./cs setup --env --jvm 21 --apps coursier)"
+       |    - sbt "domain/test; worker/test"
+       |
+       |""".stripMargin
+
   private lazy val workerTestAppVsCode   =
     s"""|// DO NOT ADJUST. This file is replaced by `./helper.scala update`.
         |{